FBSD: reduce global lock frobnication in afs_root
This change reorganizes afs_root slightly so that a temporary GLOCK drop
is unnecessary. (It was being dropped to do a vget, but this can happen
after the lock is normally released.)

Note that it should be possible to do a vget with the GLOCK held if the
GLOCK is sx(9)-based.