FBSD: support kernels built with VIMAGE
VIMAGE is FreeBSD's network virtualization support. It is not enabled by
default as of 11-RELEASE.

The main issue with VIMAGE is that without special care 'curvnet'
returns NULL, resulting in a page fault when evaluating V_xxx macros.

This support has been tested only lightly, and only for the case where
VNET support is enabled but only one "vnet" exists.

For more, see: https://wiki.freebsd.org/VIMAGE/porting-to-vimage