FBSD: only use LOCKPARENT&ISLASTCN on FreeBSD<7
Based on trial and error, attempting to keep the parent directory vnode
locked when flags LOCKPARENT and ISLASTCN are set seems to result in
deadlocks/asserts on modern FreeBSD. One path in afs_vop_lookup only
attempts to do this if the version of FreeBSD is older than 7. This
patch changes the other possible path to do the same, resulting in
improved stability.

Unfortunately, the VOP_LOOKUP(9) man page for FreeBSD 11 explicitly
claims that LOCKPARENT and ISLASTCN should result in the parent vnode
remaining locked, so I'm not sure what the correct solution is.

Change-Id: Ibac92c83b0a6e029d605ec11f5e8f52289cf0b67