FBSD: skip vinvalbuf when paging or vnode unlocked
Locking the vnode in afs_GetVCache is very problematic: whether we drop
or keep the GLOCK to do it, deadlocks can occur.

Nor does it seem feasible to insist that we always have a locked vp at
this point; it may have just been created, or the calling FBSD osi code
may not have had even known the vp yet. (E.g., when using "afs_lookup".)

Furthermore, if paging is in progress, then we will certainly deadlock
in vinvalbuf: we are the pager, so we can't wait for the pager. I have
only seen this as a result of mmap.

For now, only attempt the vinvalbuf when we already have the vp locked
and paging is not in progress.

Change-Id: I4879bb6bbede47d2f14d2c75dc196be4213e7314