FBSD: don't drop locks in osi_TryEvictVCache
Previously, the FBSD client's osi_TryEvictVCache briefly dropped the
GLOCK and afs_xvcache, presumably to avoid sleeping while holding the
GLOCK. This allows races which can result in a failed afs_vop_reclaim
after vdrop, which can lead to a panic.

This change eliminates the lock drops, since the new sx-based GLOCK can
be held by a sleeping thread. Additionally, we try to lock the vnode
without sleeping, and report failure if unable to lock immediately. This
avoids potential deadlock, since locking the vnode after the GLOCK is a
lock order reversal.

Change-Id: I9e20eadbd2b5578932b1928b5a5a8fc104055c31