FBSD: drop GLOCK for unmount's vflush
The vflush call results in several calls back into AFS which do not
expect the global lock to be held already. For example, if we call
vflush with the global lock held, we will surely panic when vflush calls
afs_root and tries to recursively obtain the global lock.

In order to have any chance at unmounting, temporarily unlock around
vflush and tolerate the consequences of the race.

Change-Id: I326f7b1cb9db2175c5938dc861238667d9b4e451