FBSD: Drop tvc->lock before VOP_LOCK for vinvalbuf
Currently, in afs_GetVCache(), we call vinvalbuf() (or the equivalent)
on BSD-ish platforms. And in order to call vinvalbuf(), we must lock
the vnode if it's not already locked by the caller. But we do this
after locking tvc->lock, so our sequence of calls can look like this:

    ObtainWriteLock(&tvc->lock, 54);
    AFS_GUNLOCK();
    vn_lock(vp, LK_EXCLUSIVE | LK_RETRY);
    vinvalbuf(vp, ...);
    VOP_UNLOCK(vp, 0);
    AFS_GLOCK();

This means we lock VOP_LOCK/vn_lock while holding tvc->lock. But in
other places in the code, we acquire tvc->lock while holding
VOP_LOCK/vn_lock, so this can result in a deadlock if we hit those
scenarios at the same time.

For example, one way we can try to acquire tvc->lock while VOP_LOCK is
already held is here:

    [...]
    #5  0xffffffff828d5af3 in afs_osi_Sleep (...) at .../src/afs/FBSD/osi_sleep.c:128
    #6  0xffffffff82847569 in Afs_Lock_Obtain (...) at .../src/afs/afs_lock.c:149
    #7  0xffffffff8285a703 in afs_VerifyVCache2 (...) at .../src/afs/afs_vcache.c:1242
    #8  0xffffffff8286d6c3 in afs_lookup (...) at .../src/afs/VNOPS/afs_vnop_lookup.c:1426
    #9  0xffffffff828d6fa6 in afs_vop_lookup (...) at .../src/afs/FBSD/osi_vnodeops.c:250
    #10 0xffffffff811f719c in VOP_LOOKUP_APV (...) at vnode_if.c:127
    #11 0xffffffff80c70041 in VOP_LOOKUP (...) at ./vnode_if.h:54
    #12 lookup (...) at /usr/src/sys/kern/vfs_lookup.c:880
    #13 0xffffffff80c6f599 in namei (...) at /usr/src/sys/kern/vfs_lookup.c:445
    #14 0xffffffff80c864e2 in kern_statat (...) at /usr/src/sys/kern/vfs_syscalls.c:2300
    #15 0xffffffff80c86c6f in sys_fstatat (...) at /usr/src/sys/kern/vfs_syscalls.c:2277
    #16 0xffffffff81074581 in syscallenter (...) at /usr/src/sys/amd64/amd64/../../kern/subr_syscall.c:135
    #17 amd64_syscall (...) at /usr/src/sys/amd64/amd64/trap.c:1186

And while holding tvc->lock, we can try to acquire VOP_LOCK inside
afs_GetVCache around the vinvalbuf call via:

    [...]
    #4  0xffffffff80b83e6e in sleeplk (...) at /usr/src/sys/kern/kern_lock.c:301
    #5  0xffffffff80b82734 in lockmgr_xlock_hard (...) at /usr/src/sys/kern/kern_lock.c:844
    #6  0xffffffff811fa9f6 in VOP_LOCK1_APV (...) at vnode_if.c:2087
    #7  0xffffffff80c8c705 in VOP_LOCK1 (...) at ./vnode_if.h:859
    #8  _vn_lock (...) at /usr/src/sys/kern/vfs_vnops.c:1548
    #9  0xffffffff8285aa9e in afs_GetVCache (...) at .../src/afs/afs_vcache.c:1755
    #10 0xffffffff828d3335 in afs_root (...) at .../src/afs/FBSD/osi_vfsops.c:213
    #11 0xffffffff80c70296 in lookup (...) at /usr/src/sys/kern/vfs_lookup.c:962
    #12 0xffffffff80c6f599 in namei (...) at /usr/src/sys/kern/vfs_lookup.c:445
    #13 0xffffffff80c864e2 in kern_statat (...) at /usr/src/sys/kern/vfs_syscalls.c:2300
    #14 0xffffffff80c86c6f in sys_fstatat (...) at /usr/src/sys/kern/vfs_syscalls.c:2277
    #15 0xffffffff81074581 in syscallenter (...) at /usr/src/sys/amd64/amd64/../../kern/subr_syscall.c:135
    #16 amd64_syscall (...) at /usr/src/sys/amd64/amd64/trap.c:1186

To fix this, we can just drop tvc->lock around the vinvalbuf() calls.
We already do this for DARWIN, so just do this for all of the
platforms where we make this call. Re-check for CStatd afterwards,
just to easily avoid unnecessary extra invalidation if the vcache
became up to date while we were waiting.

Change-Id: Iea18441aae94821188a179a9d576eb6c8f337353