Remove pattern rules not understood by some 'make' variants
The 'make' in FreeBSD's base system does not support even parsing these
rules. This causes the build to fail when make encounters, e.g.,
Makefile.lwp, which uses such a rule.

The use of these pattern rules was introduced in commit 5c64df86, but
did not cause problems on FreeBSD until commit 192a2ff4 (dir: make dtest
buildable again), as src/dir/test/Makefile is the only one in a FreeBSD
build to include Makefile.lwp.

Note that while GNU make does support such rules and is available on
FreeBSD, it's not part of the FreeBSD base system and is not supported
by FreeBSD's bsd.kmod.mk framework, which OpenAFS requires to build the
client's kernel module.

This reverts commit 5c64df8628e60f2e4c55aaf3e53fb01c140b42ed.

Change-Id: I37751d92534b99bdea06b2f7718cadd3bca954b9