Remove pattern rules not understood by some 'make' variants
The 'make' in FreeBSD's base system does not interpret these rules as
expected. This causes the build to fail when make encounters, e.g.,
src/config/Makefile.lwp, which uses such a rule:

    make[4]: don't know how to make %.c. Stop

The use of these pattern rules was introduced in commit 5c64df86, but
did not cause problems on FreeBSD until commit 192a2ff4 (dir: make dtest
buildable again), as src/dir/test/Makefile is the only one in a FreeBSD
build to include Makefile.lwp.

Note that while GNU make does support pattern rules and is available on
FreeBSD, it's not supported by FreeBSD's bsd.kmod.mk framework, which
OpenAFS requires to build the client's kernel module.

A possible issue with this commit is that the pattern rules were
originally introduced because particular 'make' versions did not
understand the equivalent double-suffix rules; hopefully this is no
longer the case. I've spot-tested on modern-day OmniOSCE, FreeBSD, and
Debian, and am hoping that BuildBot can help me to test on Solaris
10/11.

This reverts commit 5c64df8628e60f2e4c55aaf3e53fb01c140b42ed.

Change-Id: I37751d92534b99bdea06b2f7718cadd3bca954b9