FBSD: implement one-group PAGs
The FreeBSD client currently doesn't implement PAGs. On FreeBSD, PAGs are
particularly helpful for distinguishing between users in different jails which
happen to have the same numerical UID. In addition to the usual security
benefits, PAGs allow such users to authenticate individually.

Solaris seems to have a similar kernel interfaces for getting/setting groups,
including the behavior where supplementary groups are sorted. Consequentially,
this patch roughly follows the Solaris implementation.

Note that if a token not obtained in the context of a PAG, a jailed user with a
matching UID will still have the un-jailed user's token attributed to them.