FBSD: warn when asked to osi_vnhold a freed vcache
This is the wrong thing to do, we might still panic immediately anyway,
and the resulting state might be even more confusing. In the event that
we don't panic maybe I can get a sense of how often this happens and
when.

The comments in this change contain a detailed description of how we can
get to this state.

Ultimately, this is another case where temporarily dropping the global
lock causes problems. (As one might suspect.)

Change-Id: Ic0e23de124835bb2a49661ef6c5cc0664e0b5243