FBSD: disable pioctl vrele on 14+
This is a lazy attempt to address a consistent v_usecount<0 wraparound
issue. (Just `ls $file && fs whereis $file` causes it.) This seems to
appear only with newer FreeBSD (14+?). Unclear if the wraparound has
been ocurring for for some time and is only now caught by FreeBSD, or if
changes in FreeBSD are somehow resulting in a new wraparound.

We had previously removed the corresponding VN_VHOLD in 2bbba424a. The
commit message there mentions that the whole unlocking/AFS_RELE
arrangement is suspicious. With this removed the wraparound does not
seem to occur, and I also don't observe leaked references preventing
clean shutdown. Probably more careful reading of the code should be done
to decide if this is really the right thing to do.

Change-Id: I2105a9fd5fa9465dc1ee202f732e2759d7dc0f72